requirements.txt
jupyterlab
matplotlib
seaborn
notebook
ipywidgets
wandb
sklearn
hydra-core 

# TODO: freeze
